# projects/party-in-a-box/docs/api/resident.openapi.yaml
openapi: 3.1.0
info:
  title: Party in a Box — Resident API
  version: 0.1.0
  summary: Public and resident (organizer) endpoints behind the Resident screens.
  description: |
    Contract for the Resident screens on the Party in a Box Screens canvas
    (Info, New event, Petition, My events) and the public petition sign page.
    Derived from `docs/backend-spec.md` §2, §3.1–3.2, §4 and §7. Vendor and Village
    endpoints are out of scope for this file.

    Conventions (spec §3):
    - Dates are local `YYYY-MM-DD` strings in America/Chicago (never UTC timestamps).
      Timestamps are RFC 3339 UTC.
    - Every mutation accepts an `Idempotency-Key` header.
    - Errors use the `Error` shape: `{code, message, reasons[]}`.
    - The browser may preview rule results with the shared `rules` library, but the
      server's results are authoritative.

    Privacy (spec §7): public petition responses never include signer names. Only the
    organizer of a request sees its signer list. No endpoint here exposes another
    resident's contact details.
servers:
  - url: https://api.example.org/v1
    description: Placeholder until hosting is chosen
...
paths:
  /rules:
    get:
      tags: [Public]
      operationId: getRules
      security: []
      summary: Block party rules for a year
      description: Used by Info and New event. `confirmed_by_village` is false until Public Works confirms that year's rules.
      parameters:
        - name: year
          in: query
          required: false
          schema: { type: integer, example: 2027 }
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rules' }
        '404': { $ref: '#/components/responses/NotFound' }
  /blocks/lookup:
    get:
      tags: [Public]
      operationId: lookupBlock
      security: []
      summary: Find the block for an address (findBlock)
      description: New event, "Check block". Ineligible blocks include the nearest eligible block.
      parameters:
        - name: address
          in: query
          required: true
          schema: { type: string, maxLength: 120, example: 1100 S Cuyler Ave }
      responses:
        '200':
          description: Block found
          content:
            application/json:
              schema:
                type: object
                required: [block, number]
                properties:
                  block: { $ref: '#/components/schemas/BlockSummary' }
                  number: { type: integer, example: 1100 }
        '404':
          description: Address not found in Oak Park
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: not_found, message: "We couldn't find that address in Oak Park.", reasons: [] }
  /blocks/{blockId}:
    get:
      tags: [Public]
      operationId: getBlock
      security: []
      summary: Block packet with geometry
      parameters:
        - $ref: '#/components/parameters/BlockId'
      responses:
        '200':
          description: Block
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Block' }
        '404': { $ref: '#/components/responses/NotFound' }
  /checks/dates:
    post:
      tags: [Public]
      operationId: checkDates
      security: []
      summary: Live date checks for New event
      description: Petition due date, season check and the block's yearly count. The same logic runs client-side for instant feedback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [block_id, date_start, date_end]
              properties:
                block_id: { $ref: '#/components/schemas/BlockIdValue' }
                date_start: { $ref: '#/components/schemas/LocalDate' }
                date_end: { $ref: '#/components/schemas/LocalDate' }
      responses:
        '200':
          description: Check results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DateChecks' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /petitions/{token}:
    get:
      tags: [Public, Petition]
      operationId: getPetition
      security: []
      summary: What neighbors see on the petition page
      description: No signer names or contact details are ever returned here.
      parameters:
        - $ref: '#/components/parameters/PetitionToken'
      responses:
        '200':
          description: Petition
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicPetition' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: Petition closed (request submitted, withdrawn, rejected or cancelled)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /petitions/{token}/signatures:
    post:
      tags: [Public, Petition]
      operationId: signPetition
      security: []
      summary: Sign a petition (no account)
      description: |
        Counts only if the house number is on the block (spec §4.2). A second signature
        from the same address is stored as `duplicate_address` and not counted again.
      parameters:
        - $ref: '#/components/parameters/PetitionToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, house_number, consent, captcha]
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
                house_number: { type: string, pattern: '^[0-9]{1,5}( ?1/2)?$', example: '1105' }
                email: { type: string, format: email, description: Optional in MVP; used for verification in Phase 2. }
                consent: { type: boolean, const: true }
                captcha: { type: string, description: Captcha token. }
      responses:
        '201':
          description: Signature recorded
          content:
            application/json:
              schema:
                type: object
                required: [state, distinct_count, needed]
                properties:
                  state: { $ref: '#/components/schemas/SignatureState' }
                  distinct_count: { type: integer, example: 10 }
                  needed: { type: integer, example: 10 }
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: Petition closed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    residentAuth:
      type: http
      scheme: bearer
      description: Session token from passwordless email sign-in (magic link). Issued by the auth provider; not part of this contract.
  parameters:
    BlockId:
      name: blockId
      in: path
      required: true
      description: URL-encoded block id, e.g. `S%20CUYLER%20AVE%7C1100`.
      schema: { $ref: '#/components/schemas/BlockIdValue' }
...
    PetitionToken:
      name: token
      in: path
      required: true
      description: Unguessable 128-bit token (32 hex characters).
      schema: { type: string, pattern: '^[0-9a-f]{32}$' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 64 }
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unprocessable:
      description: Validation or rule check failed; `reasons` lists each one
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message, reasons]
      properties:
        code: { type: string, example: submit_blocked }
        message: { type: string }
        reasons:
          type: array
          items: { type: string }
    LocalDate:
      type: string
      format: date
      description: Local date in America/Chicago, `YYYY-MM-DD`.
      example: '2027-06-12'
    BlockIdValue:
      type: string
      description: '`<street name>|<hundred>`'
      example: S CUYLER AVE|1100
    Rules:
      type: object
      required: [rules_year, confirmed_by_village, season, hours, petition_min_addresses, petition_lead_days, max_events_per_block_per_year, max_events_per_weekend, contact, checklist]
      properties:
        rules_year: { type: integer, example: 2026 }
        confirmed_by_village: { type: boolean }
        source_url: { type: string, format: uri }
        season:
          type: object
          properties:
            start_mmdd: { type: string, example: '04-04' }
            end_mmdd: { type: string, example: '10-31' }
        hours:
          type: object
          properties:
            start: { type: string, example: '09:00' }
            end: { type: string, example: '23:00' }
        petition_min_addresses: { type: integer, example: 10 }
        petition_lead_days: { type: integer, example: 14 }
        max_events_per_block_per_year: { type: integer, example: 2 }
        max_events_per_weekend: { type: integer, example: 30 }
        contact:
          type: object
          properties:
            email: { type: string, format: email }
            phone: { type: string }
        checklist:
          type: array
          items: { type: string }
    BlockSummary:
      type: object
      required: [id, label, eligible, zip]
      properties:
        id: { $ref: '#/components/schemas/BlockIdValue' }
        label: { type: string, example: 1100 S Cuyler Ave }
        eligible: { type: boolean }
        reason: { type: string, description: Empty when eligible. }
        nearest_eligible:
          oneOf:
            - type: 'null'
            - type: object
              required: [id, label, meters]
              properties:
                id: { $ref: '#/components/schemas/BlockIdValue' }
                label: { type: string, example: 1100 S Cuyler Ave }
                meters: { type: integer, example: 54 }
        zip: { type: ['string', 'null'], example: '60304', description: 'Null until the block data has zip codes (backend-spec §5).' }
        tree_count: { type: integer, example: 55 }
        big_tree_count: { type: integer, example: 18 }
        parking:
          type: array
          items: { type: string }
        overnight_ban: { type: boolean }
    Block:
      allOf:
        - $ref: '#/components/schemas/BlockSummary'
        - type: object
          properties:
            centroid:
              type: array
              description: '[lat, lon]'
              items: { type: number }
              minItems: 2
              maxItems: 2
            lines:
              type: array
              description: Lines of [lat, lon] points.
              items:
                type: array
                items:
                  type: array
                  items: { type: number }
            data_version: { type: string }
    DateChecks:
      type: object
      required: [petition_due, due_passed, season, block_year_count, max_per_block_per_year]
      properties:
        petition_due: { $ref: '#/components/schemas/LocalDate' }
        due_passed: { type: boolean }
        season:
          type: object
          required: [ok, problems]
          properties:
            ok: { type: boolean }
            problems:
              type: array
              items: { type: string }
              example: ['2027-11-02 is outside the Apr 4–Oct 31 season']
        block_year_count: { type: integer, example: 0 }
        max_per_block_per_year: { type: integer, example: 2 }
...
    PublicPetition:
      type: object
      required: [block_label, date_start, date_end, hours, organizer_display_name, distinct_count, needed, open]
      properties:
        block_label: { type: string, example: 1100 S Cuyler Ave }
        date_start: { $ref: '#/components/schemas/LocalDate' }
        date_end: { $ref: '#/components/schemas/LocalDate' }
        hours: { type: string, example: 9 a.m. – 11 p.m. }
        barricades: { type: boolean }
        organizer_display_name: { type: string, example: Sample Organizer }
        distinct_count: { type: integer, example: 9 }
        needed: { type: integer, example: 10 }
        open: { type: boolean }
    SignatureState:
      type: string
      enum: [counted, duplicate_address, off_block, struck]
...